Bozza funzionante menù card chiamate
