+++ lib/int/ajax_create.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once '../init.php';
+require_once '../../init.php';
 
 session_start();
 if (!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true) {
